undo the code changes
