+++ ons_alpha/core/blocks/stream_blocks.py
@@ -1,4 +1,5 @@
-from wagtail.blocks import RichTextBlock, StreamBlock
+from django.conf import settings
+from wagtail.blocks import CharBlock, RichTextBlock, StreamBlock, StructBlock
 from wagtail.embeds.blocks import EmbedBlock
 from wagtail.images.blocks import ImageChooserBlock
 from wagtailmath.blocks import MathBlock
@@ -15,11 +16,26 @@ from ons_alpha.core.blocks import (
     RelatedLinksBlock,
 )
 from ons_alpha.core.blocks import (
-    TableBlock as OldTableBlock,  # Retain this for other usages in stream_blocks
+    TableBlock as OldTableBlock,  # Custom TableBlock for existing data
 )
-from ons_alpha.core.blocks.markup import ONSTableBlock  # Import the relocated ONSTableBlock
 
 
+# StructBlock for ONS Table
+class ONSTableBlock(StructBlock):
+    heading = CharBlock(required=True, help_text="Add a heading for the table.")
+    table = OldTableBlock(required=True, help_text="Add the table data here.")
+    source = CharBlock(required=False, help_text="Add the source of the table data if applicable.")
+    footnotes = RichTextBlock(
+        features=settings.RICH_TEXT_BASIC, required=False, help_text="Add any footnotes for the table."
+    )
+
+    class Meta:
+        template = "templates/components/streamfield/ons_table_block.html"
+        icon = "table"
+        label = "ONS Table"
+
+
+# Core story block with added ONSTableBlock
 class CoreStoryBlock(StreamBlock):
     heading = HeadingBlock()
     rich_text = RichTextBlock()
@@ -38,6 +54,7 @@ class CoreStoryBlock(StreamBlock):
         block_counts = {"related_links": {"max_num": 1}}
 
 
+# Block for corrections and notices
 class CorrectionsNoticesStoryBlock(StreamBlock):
     correction = CorrectionBlock()
     notice = NoticeBlock()
