+++ ons_alpha/release_calendar/models.py
@@ -15,18 +15,20 @@ from ons_alpha.release_calendar.blocks import ReleaseStoryBlock
 from ons_alpha.utils.models import LinkFields
 
 
+# Status choices for ReleasePage
 class ReleaseStatus(models.TextChoices):
     PROVISIONAL = _("provisional"), _("Provisional")
     CONFIRMED = _("confirmed"), _("Confirmed")
     CANCELLED = _("cancelled"), _("Cancelled")
 
 
+# Model for ReleaseIndex
 class ReleaseIndex(BasePage):
     template = "templates/pages/release_index.html"
 
     parent_page_types = ["home.HomePage"]
     subpage_types = ["ReleasePage"]
-    max_count_per_parent = 1
+    max_count_per_parent = 1  # Set max count per parent
 
     def get_context(self, request, *args, **kwargs):
         context = super().get_context(request, *args, **kwargs)
@@ -41,10 +43,16 @@ class ReleaseIndex(BasePage):
         return context
 
 
+# Model for related links in ReleasePage
 class ReleasePageRelatedLink(Orderable, LinkFields):
+    """
+    Related links. e.g. https://www.ons.gov.uk/releases/welshlanguagecensus2021inwales
+    """
+
     parent = ParentalKey("ReleasePage", related_name="related_links", on_delete=models.CASCADE)
 
 
+# Model for ReleasePage
 class ReleasePage(BasePage):
     template = "templates/pages/release_page.html"
 
@@ -83,7 +91,7 @@ class ReleasePage(BasePage):
         MultiFieldPanel(
             [
                 FieldRowPanel(
-                    [FieldPanel("release_calendar_page"), FieldPanel("publication_date")],
+                    [FieldPanel("release_date"), FieldPanel("next_release")],
                     heading="Dates",
                 ),
                 FieldPanel("status"),
@@ -127,7 +135,8 @@ class ReleasePage(BasePage):
         context["related_links"] = self.related_links_for_context
         context["toc"] = self.toc
 
-        for block in self.content:
+        # Suppress the Pylint warning since we know self.content is iterable
+        for block in self.content:  # pylint: disable=not-an-iterable
             context["toc"] += block.block.to_table_of_contents_items(block.value)
 
         return context
@@ -147,7 +156,7 @@ class ReleasePage(BasePage):
         items = [{"url": "#summary", "text": _("Summary")}]
 
         if self.status == ReleaseStatus.PUBLISHED:
-            for block in self.content:
+            for block in self.content:  # pylint: disable=not-an-iterable
                 items += block.block.to_table_of_contents_items(block.value)
 
             if self.datasets:
